Extract a common base for different test frameworks
