+++ testsupport/common/pom.xml
@@ -28,19 +28,19 @@
         <relativePath>../..</relativePath>
     </parent>
 
-    <artifactId>s3mock-junit4</artifactId>
+    <artifactId>s3mock-testsupport-common</artifactId>
     <packaging>jar</packaging>
 
-    <name>S3Mock - Testsupport - JUnit4</name>
+    <name>S3Mock - Testsupport - Common</name>
 
     <dependencies>
         <dependency>
             <groupId>com.adobe.testing</groupId>
-            <artifactId>s3mock-testsupport-common</artifactId>
+            <artifactId>s3mock</artifactId>
         </dependency>
         <dependency>
-            <groupId>junit</groupId>
-            <artifactId>junit</artifactId>
+            <groupId>com.amazonaws</groupId>
+            <artifactId>aws-java-sdk-s3</artifactId>
         </dependency>
     </dependencies>
 </project>
