+++ testsupport/junit4/src/main/java/com/adobe/testing/s3mock/junit4/S3MockRule.java
@@ -16,30 +16,12 @@
 
 package com.adobe.testing.s3mock.junit4;
 
-import static java.lang.String.join;
-
-import com.adobe.testing.s3mock.S3MockApplication;
-import com.amazonaws.ClientConfiguration;
-import com.amazonaws.auth.AWSStaticCredentialsProvider;
-import com.amazonaws.auth.BasicAWSCredentials;
-import com.amazonaws.client.builder.AwsClientBuilder.EndpointConfiguration;
-import com.amazonaws.services.s3.AmazonS3;
-import com.amazonaws.services.s3.AmazonS3ClientBuilder;
-import java.net.Socket;
-import java.security.KeyManagementException;
-import java.security.NoSuchAlgorithmException;
-import java.security.cert.CertificateException;
-import java.security.cert.X509Certificate;
-import java.util.HashMap;
+import com.adobe.testing.s3mock.testsupport.common.S3MockStarter;
 import java.util.Map;
-import javax.net.ssl.SSLContext;
-import javax.net.ssl.SSLEngine;
-import javax.net.ssl.TrustManager;
-import javax.net.ssl.X509ExtendedTrustManager;
-import org.apache.http.conn.ssl.NoopHostnameVerifier;
-import org.apache.http.conn.ssl.SSLConnectionSocketFactory;
 import org.junit.ClassRule;
-import org.junit.rules.ExternalResource;
+import org.junit.rules.TestRule;
+import org.junit.runner.Description;
+import org.junit.runners.model.Statement;
 
 /**
  * JUnit rule to start and stop the S3Mock Application. After the
@@ -57,12 +39,12 @@ import org.junit.rules.ExternalResource;
  * }
  * </pre>
  */
-public class S3MockRule extends ExternalResource {
-  private S3MockApplication s3MockFileStore;
-  private final Map<String, Object> properties;
-
+public class S3MockRule extends S3MockStarter implements TestRule {
+  /**
+   * Creates an instance with the default configuration.
+   */
   public S3MockRule() {
-    properties = defaultProps();
+    super(null);
   }
 
   /**
@@ -72,194 +54,27 @@ public class S3MockRule extends ExternalResource {
     return new Builder();
   }
 
-  S3MockRule(final Map<String, Object> properties) {
-    this();
-    this.properties.putAll(properties);
-  }
-
-  private static Map<String, Object> defaultProps() {
-    final Map<String, Object> args = new HashMap<>();
-    args.put(S3MockApplication.PROP_HTTPS_PORT, "0");
-    args.put(S3MockApplication.PROP_HTTP_PORT, "0");
-    return args;
-  }
-
-  /**
-   * @return An {@link AmazonS3} client instance that is configured to call the started S3Mock
-   *         server using HTTPS.
-   */
-  public AmazonS3 createS3Client() {
-    final BasicAWSCredentials credentials = new BasicAWSCredentials("foo", "bar");
-
-    return AmazonS3ClientBuilder.standard()
-        .withCredentials(new AWSStaticCredentialsProvider(credentials))
-        .withClientConfiguration(
-            configureClientToIgnoreInvalidSslCertificates(new ClientConfiguration()))
-        .withEndpointConfiguration(
-            new EndpointConfiguration("https://localhost:" + getPort(), "us-east-1"))
-        .enablePathStyleAccess()
-        .build();
-  }
-
-  /**
-   * @return The HTTPS port that the S3Mock uses.
-   */
-  public int getPort() {
-    return s3MockFileStore.getPort();
-  }
-
-  /**
-   * @return The HTTP port that the S3Mock uses.
-   */
-  public int getHttpPort() {
-    return s3MockFileStore.getHttpPort();
-  }
-
-  /**
-   * Registers a valid KMS key reference in the mock server.
-   *
-   * @param keyRef A KMS Key Reference
-   */
-  public void registerKMSKeyRef(final String keyRef) {
-    s3MockFileStore.registerKMSKeyRef(keyRef);
-  }
-
-  /**
-   * Adjusts the given client configuration to allow the communication with the mock server using
-   * HTTPS, although that one uses a self-signed SSL certificate.
-   *
-   * @param clientConfiguration The {@link ClientConfiguration} to adjust.
-   * @return The adjusted instance.
-   */
-  public ClientConfiguration configureClientToIgnoreInvalidSslCertificates(
-      final ClientConfiguration clientConfiguration) {
-
-    clientConfiguration.getApacheHttpClientConfig()
-        .withSslSocketFactory(new SSLConnectionSocketFactory(
-            createBlindlyTrustingSSLContext(),
-            NoopHostnameVerifier.INSTANCE));
-
-    return clientConfiguration;
+  private S3MockRule(final Map<String, Object> properties) {
+    super(properties);
   }
 
   @Override
-  protected void before() {
-    s3MockFileStore = S3MockApplication.start(properties);
-  }
-
-  @Override
-  protected void after() {
-    s3MockFileStore.stop();
-  }
-
-  private SSLContext createBlindlyTrustingSSLContext() {
-    try {
-      final SSLContext sc = SSLContext.getInstance("TLS");
-
-      sc.init(null, new TrustManager[] {new X509ExtendedTrustManager() {
-        @Override
-        public java.security.cert.X509Certificate[] getAcceptedIssuers() {
-          return null;
-        }
-
-        @Override
-        public void checkClientTrusted(final X509Certificate[] certs, final String authType) {
-          // no-op
-        }
-
-        @Override
-        public void checkServerTrusted(final X509Certificate[] certs, final String authType) {
-          // no-op
-        }
-
-        @Override
-        public void checkClientTrusted(final X509Certificate[] arg0, final String arg1,
-            final Socket arg2)
-            throws CertificateException {
-          // no-op
-        }
-
-        @Override
-        public void checkClientTrusted(final X509Certificate[] arg0, final String arg1,
-            final SSLEngine arg2)
-            throws CertificateException {
-          // no-op
-        }
-
-        @Override
-        public void checkServerTrusted(final X509Certificate[] arg0, final String arg1,
-            final Socket arg2)
-            throws CertificateException {
-          // no-op
-        }
-
-        @Override
-        public void checkServerTrusted(final X509Certificate[] arg0, final String arg1,
-            final SSLEngine arg2)
-            throws CertificateException {
-          // no-op
+  public Statement apply(final Statement base, final Description description) {
+    return new Statement() {
+      @Override
+      public void evaluate() throws Throwable {
+        start();
+        try {
+          base.evaluate();
+        } finally {
+          stop();
         }
       }
-      }, new java.security.SecureRandom());
-
-      return sc;
-    } catch (final NoSuchAlgorithmException | KeyManagementException e) {
-      throw new RuntimeException("Unexpected exception", e);
-    }
+    };
   }
 
-  public static class Builder {
-    private final Map<String, Object> arguments = new HashMap<>();
-
-    /**
-     * @param initialBuckets buckets that are to be created at startup
-     * @return the builder
-     */
-    public Builder withInitialBuckets(final String... initialBuckets) {
-      arguments.put(S3MockApplication.PROP_INITIAL_BUCKETS, join(",", initialBuckets));
-      return this;
-    }
-
-    /**
-     * @param httpsPort The HTTPS port to use. If not configured, a random port will be used.
-     * @return the builder
-     */
-    public Builder withHttpsPort(final int httpsPort) {
-      arguments.put(S3MockApplication.PROP_HTTPS_PORT, String.valueOf(httpsPort));
-      return this;
-    }
-
-    /**
-     * @param httpPort The HTTP port to use. If not configured, a random port will be used.
-     * @return the builder
-     */
-    public Builder withHttpPort(final int httpPort) {
-      arguments.put(S3MockApplication.PROP_HTTP_PORT, String.valueOf(httpPort));
-      return this;
-    }
-
-    /**
-     * @param rootFolder The root directory to use. If not configured, a default temp-dir will be
-     *                  used.
-     * @return the builder
-     */
-    public Builder withRootFolder(final String rootFolder) {
-      arguments.put(S3MockApplication.PROP_ROOT_DIRECTORY, rootFolder);
-      return this;
-    }
-
-    /**
-     * Reduces logging level WARN and suppresses the startup banner.
-     * @return the builder
-     */
-    public Builder silent() {
-      arguments.put(S3MockApplication.PROP_SILENT, true);
-      return this;
-    }
-
-    /**
-     * @return The configured instance.
-     */
+  public static class Builder extends S3MockStarter.BaseBuilder<S3MockRule> {
+    @Override
     public S3MockRule build() {
       return new S3MockRule(arguments);
     }
